chore(shared): migrate `expressionTypeMap` and related types to `@repo/shared`
