+++ apps/admin/package.json
@@ -31,7 +31,8 @@
     "tailwind-merge": "^3.3.1",
     "usehooks-ts": "^3.1.1",
     "uuid": "^9.0.1",
-    "zod": "^4.1.11",
+    "zod": "^4.3.6",
+    "@repo/shared": "*",
     "@repo/ui": "*"
   },
   "devDependencies": {
