+++ apps/app/package.json
@@ -33,7 +33,8 @@
     "sonner": "^2.0.7",
     "tailwind-merge": "^3.3.1",
     "usehooks-ts": "^3.1.1",
-    "zod": "^4.1.12",
+    "zod": "^4.3.6",
+    "@repo/shared": "*",
     "@repo/ui": "*"
   },
   "devDependencies": {
