+++ packages/shared/src/types.ts
@@ -1,0 +1,13 @@
+import { z } from "zod";
+
+export const expressionContextTypeSchema = z.enum([
+  "verb",
+  "noun",
+  "adjective",
+  "adverb",
+  "phrasal_verb",
+  "irregular_verb",
+  "simple_expression",
+]);
+
+export type ExpressionContextType = z.infer<typeof expressionContextTypeSchema>;
